Fix for #572: merge ~/storage instead of del
We could recursively delete, as in the method above, but that would
require a separate try block for a separate logcat message, or a
different method that returns a flag. I prefer merging in case of custom
folders. Another alternative may be to send an exit code or something for
the termux-setup-storage script to detect and echo an error message.