Refactor: Extract `getIndexWithDirection()`, `isSessionPendingResult()` from `TermuxTerminalSessionClient.java`
In function `onSessionFinished()`, When a boolean variable
`isPluginExecutionCommandWithPendingResult` is
assigned, the process is included in a function as it is, which reduces
readability. So extract and improve it.
In function `switchToSession()`, When `Index` is assigned,
the process is included in a function as it is, which reduces
readability. So extract and improve it.