Added: TerminalView accessibility actions to announce cursor position
There's a bit of a discrepancy where the "accessibility" cursor is, and where
the true terminal cursor is. So add a couple of actions to help figure out
where your text will go, or what it is you've just typed. Without this, there
is currently nothing on the Android UI side to tell you where it is.